product bugs updated

+++ pkg/repository/produtRepo.go
@@ -27,12 +27,25 @@ func (c *ProductDatabase) CreateCategory(category helperStruct.Category) (respon
 
 func (c *ProductDatabase) UpdatCategory(category helperStruct.Category, id int) (response.Category, error) {
 	var updatedCategory response.Category
-	query := `UPDATE  categories SET category_name = $1 , updated_at =NOW() WHERE id=$2 RETURNING id,category_name `
+	query := `UPDATE  categories SET category_name = $1 , updated_at =NOW() WHERE EXISTS(SELECT 1 FROM categories WHERE id=$2) RETURNING id,category_name `
 	err := c.DB.Raw(query, category.Name, id).Scan(&updatedCategory).Error
-	return updatedCategory, err
+	if err != nil {
+		return response.Category{}, err
+	}
+	if updatedCategory.Id == 0 {
+		return response.Category{}, fmt.Errorf("no such category to update")
+	}
+	return updatedCategory, nil
 }
 
 func (c *ProductDatabase) DeleteCategory(id int) error {
+	var exits bool
+
+	query1 := `select exists(select 1 from categories where id=?)`
+	c.DB.Raw(query1, id).Scan(&exits)
+	if !exits {
+		return fmt.Errorf("no category found")
+	}
 	query := `DELETE FROM categories WHERE id=$1`
 	err := c.DB.Exec(query, id).Error
 	return err
@@ -47,9 +60,22 @@ func (c *ProductDatabase) ListCategories() ([]response.Category, error) {
 
 func (c *ProductDatabase) DisplayCategory(id int) (response.Category, error) {
 	var category response.Category
+	var exits bool
+
+	query1 := `select exists(select 1 from categories where id=?)`
+	c.DB.Raw(query1, id).Scan(&exits)
+	if !exits {
+		return response.Category{}, fmt.Errorf("no category found")
+	}
 	query := `SELECT * FROM categories WHERE id=$1`
 	err := c.DB.Raw(query, id).Scan(&category).Error
-	return category, err
+	if err != nil {
+		return response.Category{}, err
+	}
+	if category.Id == 0 {
+		return response.Category{}, fmt.Errorf("no such category")
+	}
+	return category, nil
 }
 
 func (c *ProductDatabase) AddProduct(product helperStruct.Product) (response.Product, error) {
@@ -83,10 +109,22 @@ func (c *ProductDatabase) UpdateProduct(id int, product helperStruct.Product) (r
 		RETURNING id,product_name,description,brand,category_id`
 	err := c.DB.Raw(query2, product.Name, product.Description, product.Brand, product.CategoryId, id).
 		Scan(&updatedProduct).Error
-	return updatedProduct, err
+	if err != nil {
+		return response.Product{}, err
+	}
+	if updatedProduct.Id == 0 {
+		return response.Product{}, fmt.Errorf("there is no such product")
+	}
+	return updatedProduct, nil
 }
 
 func (c *ProductDatabase) DeleteProduct(id int) error {
+	var exists bool
+	isExists := `SELECT EXISTS (SELECT 1 FROM products WHERE id=$1)`
+	c.DB.Raw(isExists, id).Scan(&exists)
+	if !exists {
+		return fmt.Errorf("there is no such product to delete")
+	}
 	query := `DELETE FROM products WHERE id=$1`
 	err := c.DB.Exec(query, id).Error
 	return err
@@ -135,6 +173,13 @@ func (c *ProductDatabase) AddProductItem(productItem helperStruct.ProductItem) (
 }
 
 func (c *ProductDatabase) UpdateProductItem(id int, productItem helperStruct.ProductItem) (response.ProductItem, error) {
+	var exists bool
+	isExists := `SELECT EXISTS (SELECT 1 FROM product_items WHERE id=$1)`
+	c.DB.Raw(isExists, id).Scan(&exists)
+	if !exists {
+		return response.ProductItem{}, fmt.Errorf("there is no such product to update")
+	}
+
 	var updatedItem response.ProductItem
 	query := `UPDATE product_items SET 
 	product_id=$1,
@@ -181,6 +226,12 @@ func (c *ProductDatabase) UpdateProductItem(id int, productItem helperStruct.Pro
 }
 
 func (c *ProductDatabase) DeleteProductItem(id int) error {
+	var exists bool
+	isExists := `SELECT EXISTS (SELECT 1 FROM product_items WHERE id=$1)`
+	c.DB.Raw(isExists, id).Scan(&exists)
+	if !exists {
+		return fmt.Errorf("there is no such product to delete")
+	}
 	query := `DELETE FROM product_items WHERE id=?`
 	err := c.DB.Exec(query, id).Error
 	return err
@@ -237,6 +288,9 @@ func (c *ProductDatabase) DisaplyProductItem(id int) (response.ProductItem, erro
 	if err != nil {
 		return response.ProductItem{}, err
 	}
+	if productItem.Id == 0 {
+		return response.ProductItem{}, fmt.Errorf("there is no such product item")
+	}
 	getImages := `SELECT file_name FROM images WHERE product_item_id=$1`
 	err = c.DB.Raw(getImages, id).Scan(&productItem.Image).Error
 	if err != nil {
@@ -286,6 +340,12 @@ func (c *ProductDatabase) ShowProduct(id int) (response.Product, error) {
 	query := `SELECT p.product_name,p.description,p.brand,c.category_name FROM products p 
 		JOIN categories c ON p.category_id=c.id WHERE p.id=$1`
 	err := c.DB.Raw(query, id).Scan(&product).Error
+	if err != nil {
+		return response.Product{}, err
+	}
+	if product.Id == 0 {
+		return response.Product{}, fmt.Errorf("there is no such product")
+	}
 	return product, err
 }
